% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshotButton.R
\name{snapshotButton}
\alias{snapshotButton}
\title{Button to execute the takeSnapshot() button}
\usage{
snapshotButton()
}
\description{
Button to execute the takeSnapshot() button
}
\examples{
snapshotButton()

}
\author{
Emelie Hofland, \email{emelie_hofland@hotmail.com}
}
