% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startWebcam.R
\name{startWebcam}
\alias{startWebcam}
\title{Function that starts the webcam and shows webcam output in a Shiny app}
\usage{
startWebcam(width, height, quality)
}
\arguments{
\item{width}{numeric, width of the webcam output}

\item{height}{numeric, height of the webcam output}

\item{quality}{numeric, for JPEG images, this is the desired quality, from 0 (worst) to 100 (best).}
}
\description{
Function that starts the webcam and shows webcam output in a Shiny app
}
\examples{
startWebcam(320,240,100)

}
\author{
Emelie Hofland, \email{emelie_hofland@hotmail.com}
}
