% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/takeSnapshot.R
\name{takeSnapshot}
\alias{takeSnapshot}
\title{Function takes a snapshot with the webcam in a R Shiny App}
\usage{
takeSnapshot()
}
\description{
Function takes a snapshot with the webcam in a R Shiny App
}
\examples{
takeSnapshot()

}
\author{
Emelie Hofland, \email{emelie_hofland@hotmail.com}
}
